+++ src/main/java/com/breakinblocks/plonk/client/render/tile/TESRPlacedItems.java
@@ -1,11 +1,11 @@
 package com.breakinblocks.plonk.client.render.tile;
 
+import com.breakinblocks.plonk.client.util.RenderUtils;
 import com.breakinblocks.plonk.common.block.BlockPlacedItems;
 import com.breakinblocks.plonk.common.tile.TilePlacedItems;
 import com.breakinblocks.plonk.common.tile.TilePlacedItems.ItemMeta;
 import com.breakinblocks.plonk.common.util.MatrixUtils;
 import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.platform.GlStateManager;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.*;
@@ -20,7 +20,6 @@ import net.minecraft.entity.item.ItemFrameEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.state.DirectionProperty;
 import net.minecraft.util.Direction;
-import net.minecraftforge.common.model.TransformationHelper;
 
 import static com.breakinblocks.plonk.common.tile.TilePlacedItems.RENDER_TYPE_BLOCK;
 import static com.breakinblocks.plonk.common.tile.TilePlacedItems.RENDER_TYPE_ITEM;
@@ -34,6 +33,7 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
     private static final Tessellator tessellator = Tessellator.getInstance();
     private static final BufferBuilder bufferbuilder = tessellator.getBuffer();
     private static final BlockRendererDispatcher blockRendererDispatcher = mc.getBlockRendererDispatcher();
+    private static final double EPS = 0.001;
 
     public TESRPlacedItems(TileEntityRendererDispatcher rendererDispatcherIn) {
         super(rendererDispatcherIn);
@@ -41,8 +41,8 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
 
     public static int getRenderTypeFromStack(ItemStack itemstack) {
         IBakedModel model = itemRenderer.getItemModelWithOverrides(itemstack, null, null);
-        Matrix4f matrixFixed = TransformationHelper.toTransformation(model.getItemCameraTransforms().getTransform(ItemCameraTransforms.TransformType.FIXED)).getMatrix();
-        Matrix4f matrixGui = TransformationHelper.toTransformation(model.getItemCameraTransforms().getTransform(ItemCameraTransforms.TransformType.GUI)).getMatrix();
+        Matrix4f matrixFixed = RenderUtils.getModelTransformMatrix(model, ItemCameraTransforms.TransformType.FIXED);
+        Matrix4f matrixGui = RenderUtils.getModelTransformMatrix(model, ItemCameraTransforms.TransformType.GUI);
         Matrix4f difference = MatrixUtils.difference(matrixFixed, matrixGui);
         MatrixUtils.TransformData transform = new MatrixUtils.TransformData(difference);
 
@@ -60,8 +60,12 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
         //String message = String.format("hS=%.3f hRot=%.3f\n", hS, hRot) + transform.toString();
         //for (String line : message.split("\r?\n"))
         //    Minecraft.getInstance().ingameGUI.getChatGUI().printChatMessage(new StringTextComponent(line));
-        // Change in scaling is < 1 and change in rotation at non-right angles;
-        return hS < (1.0 - 0.001) && hRot > 0.001 ? RENDER_TYPE_BLOCK : RENDER_TYPE_ITEM;
+        // The following is a heuristic
+        // Use block rendering: hRot ~= 83.80586
+        final double blockRot = 83.80586;
+        if (blockRot - EPS <= hRot && hRot <= blockRot + EPS)
+            return RENDER_TYPE_BLOCK;
+        return RENDER_TYPE_ITEM;
     }
 
     /**
@@ -120,7 +124,7 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
         matrixStackIn.translate(0.0, -0.5, 0.0);
 
         //Rotate about axis
-        matrixStackIn.rotate(new Quaternion(Vector3f.YP, (float) -tileEntityIn.getTileRotationAngle(), true));
+        matrixStackIn.rotate(Vector3f.YP.rotationDegrees((float) -tileEntityIn.getTileRotationAngle()));
 
         ItemStack[] contents = tileEntityIn.getContentsDisplay();
         ItemMeta[] contentsMeta = tileEntityIn.getContentsMeta();
@@ -175,13 +179,11 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
     public void renderStack(float partialTicks, MatrixStack matrixStackIn, IRenderTypeBuffer bufferIn, int combinedLightIn, ItemStack stack, ItemMeta meta, boolean halfSize) {
         if (stack.isEmpty()) return;
         matrixStackIn.push();
-        GlStateManager.disableLighting();
-        GlStateManager.pushLightingAttributes();
-        RenderHelper.enableStandardItemLighting();
+        //GlStateManager.disableLighting();
 
         // Rotate item
         //GL11.glRotated(-meta.getRotationAngle(), 0.0, 1.0, 0.0);
-        matrixStackIn.rotate(new Quaternion(new Vector3f(0f, 1f, 0f), (float) -meta.getRotationAngle(), true));
+        matrixStackIn.rotate(Vector3f.YP.rotationDegrees((float) -meta.getRotationAngle()));
 
         // GROUND
         //matrixStackIn.translate(0f, -0.125f, 0f);
@@ -189,7 +191,10 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
 
         // FIXED
         //GlStateManager.rotate(180f, 0f, 1f, 0f);
-        matrixStackIn.rotate(new Quaternion(new Vector3f(0f, 1f, 0f), 180, true));
+        matrixStackIn.rotate(new Quaternion(Vector3f.YP, 180, true));
+
+        //GlStateManager.pushLightingAttributes();
+        //RenderHelper.enableStandardItemLighting();
         switch (meta.renderType) {
             case RENDER_TYPE_BLOCK: {
                 matrixStackIn.translate(0f, 0.25f, 0f);
@@ -201,15 +206,15 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
             default:
                 matrixStackIn.translate(0f, 2f / 48, 0f);
                 //GlStateManager.rotate(90f, 1f, 0f, 0f);
-                matrixStackIn.rotate(new Quaternion(new Vector3f(1f, 0f, 0f), 90, true));
+                matrixStackIn.rotate(Vector3f.XP.rotationDegrees(90f));
                 if (halfSize)
                     matrixStackIn.scale(0.5F, 0.5F, 0.5F);
                 itemRenderer.renderItem(stack, ItemCameraTransforms.TransformType.FIXED, combinedLightIn, OverlayTexture.NO_OVERLAY, matrixStackIn, bufferIn);
         }
+        //RenderHelper.disableStandardItemLighting();
+        //GlStateManager.popAttributes();
 
-        RenderHelper.disableStandardItemLighting();
-        GlStateManager.popAttributes();
-        GlStateManager.enableLighting();
+        //GlStateManager.enableLighting();
         matrixStackIn.pop();
     }
 }